requirements/lmdeploy.txt
accelerate==0.33.0
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
airportsdata==20241001
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
av==13.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
datasets==3.0.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
fire==0.6.0
frozenlist==1.5.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
interegular==0.3.3
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark==1.2.2
llava @ git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd
lmdeploy==0.6.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmengine-lite==0.10.4
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai==1.52.2
outlines==0.1.1
outlines_core==0.1.14
packaging==24.1
pandas==2.2.3
peft==0.11.1
pillow==10.4.0
platformdirs==4.2.2
propcache==0.2.0
protobuf==5.27.3
psutil==6.0.0
pyarrow==17.0.0
pycountry==24.6.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pynvml==11.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
qwen-vl-utils==0.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.3
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.20.1
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers @ git+https://github.com/huggingface/transformers@1d063793318b20654ebb850f48f43e0a247ab7bb
triton==2.2.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.30.4
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
xxhash==3.5.0
yapf==0.40.2
yarl==1.16.0
zipp==3.19.2
